# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Restaurant Management System HTTP API.",
        "title": "RMS API",
        "contact": {
            "name": "RMS"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/menu-items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu-items"
                ],
                "summary": "List menu items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-array_github_com_Jiruu246_rms_internal_dto_MenuItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu-items"
                ],
                "summary": "Create a menu item",
                "parameters": [
                    {
                        "description": "Menu item details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.CreateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_MenuItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/menu-items/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu-items"
                ],
                "summary": "Get a menu item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_MenuItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu-items"
                ],
                "summary": "Update a menu item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.UpdateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_MenuItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "menu-items"
                ],
                "summary": "Delete a menu item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Jiruu246_rms_internal_dto.CreateMenuItemRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "restaurant_id"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_internal_dto.MenuItem": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                },
                "modifiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.Modifier"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "github_com_Jiruu246_rms_internal_dto.Modifier": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "max": {
                    "type": "integer"
                },
                "multi_select": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_internal_dto.UpdateMenuItemRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-array_github_com_Jiruu246_rms_internal_dto_MenuItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.MenuItem"
                    }
                },
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_MenuItem": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.MenuItem"
                },
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}